requirements-client.txt
websockets>=12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyQt6>=6.5.0
qasync>=0.24.0
